requirements.txt
py-sudoku
numpy
torch
matplotlib
tdqm

pytest
pydantic